refactor(request): extract head green evaluation
